+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Changed the MinIO delete cleanup interval from 5 minutes to 1 minute
 - Improve domains endpoint performance
 - Add a top-level DeltaFile field listing indexedMetadata keys, improve performance of indexedMetadataKeys endpoint
+- Cap total count at 50k for deltaFiles query to improve mongodb performance
+- Add totalCount endpoint to give estimated total count of deltaFiles
 
 ### Deprecated